clear repo
