+++ package.json
@@ -2,9 +2,8 @@
   "name": "react-forms-builder",
   "version": "0.1.0",
   "license": "MIT",
-  "homepage": "react-forms",
+  "homepage": "",
   "devDependencies": {
-    "chai": "^4.1.2",
     "prettier": "1.13.0",
     "react-scripts": "^1.1.4"
   },
