--- src/App.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { CompleteForm } from "./CompleteForm";
-
-export class App extends React.Component {
-    render() {
-        return (<React.Fragment>
-                <CompleteForm />
-            </React.Fragment>);
-    }
-}
